Major refactorings all around

+++ pages/api/asset/index.ts
@@ -2,24 +2,25 @@ import { createReadStream } from "fs";
 import { unlink } from "fs/promises";
 import { uuid } from "util/uuid";
 import { NextApiResponse } from "next";
-import { GetAssetHeaders, GetAssetQuery, PostAssetQuery } from "type/asset";
 import { parseRequest } from "util/parseRequest";
 import { pathToFileURL } from "url";
 import { ParsedMultipartBody, parseMultipartBody } from "util/multipart";
-import { AssetService } from "service/asset";
+import { AssetService, GetAssetHeaders } from "service/asset";
 import { RequestWithLogger, withDefaults } from "util/withDefaults";
+import { AssetData, Space } from "type/space";
+import { z } from "zod";
 
 async function asset(req: RequestWithLogger, res: NextApiResponse): Promise<void> {
   const svc = new AssetService(req);
 
-  async function post(query: PostAssetQuery, body: ParsedMultipartBody) {
+  async function post(spaceId: Space["spaceId"], body: ParsedMultipartBody) {
     const out = await Promise.all(
       Object.entries(body.files).map(async ([field, file]) => {
         const assetId = uuid();
         const s = createReadStream(file.path);
         const data = {
           assetId,
-          spaceId: query.spaceId,
+          spaceId,
           filename: file.originalFilename,
           size: file.size,
           contentType: file.headers["content-type"],
@@ -33,41 +34,40 @@ async function asset(req: RequestWithLogger, res: NextApiResponse): Promise<void
     return Object.fromEntries(out);
   }
 
-  switch (req.method) {
-    case "HEAD": {
-      const { query } = parseRequest(req, {
-        query: GetAssetQuery,
-      });
-      const r = await svc.headAsset(query);
-      r.write(res);
+  if (req.method === "HEAD") {
+    const { query } = parseRequest(req, {
+      query: z.object({ assetId: AssetData.shape.assetId }),
+    });
 
-      return;
-    }
+    const r = await svc.headAsset(query);
+    r.write(res);
+  }
 
-    case "GET": {
-      const { query, headers } = parseRequest(req, {
-        query: GetAssetQuery,
-        headers: GetAssetHeaders,
-      });
-      const r = await svc.getAsset(query, headers);
-      r.write(res);
+  if (req.method === "GET") {
+    const { query, headers } = parseRequest(req, {
+      query: z.object({ assetId: z.string().uuid() }),
+      headers: GetAssetHeaders,
+    });
 
-      return;
-    }
+    const r = await svc.getAsset(query, headers);
+    r.write(res);
+  }
 
-    case "POST": {
-      const { query } = parseRequest(req, {
-        query: PostAssetQuery,
-      });
-      const body = await parseMultipartBody(req);
-      const assetIds = await post(query, body);
-      res.status(201).json({ assetIds });
-      await Promise.all(
-        Object.values(body.files).map(({ path }) => unlink(pathToFileURL(path)))
-      );
+  if (req.method === "POST") {
+    const {
+      query: { spaceId },
+    } = parseRequest(req, {
+      query: z.object({ spaceId: Space.shape.spaceId }),
+    });
 
-      return;
-    }
+    const body = await parseMultipartBody(req);
+    const assetIds = await post(spaceId, body);
+
+    res.status(201).json({ assetIds });
+
+    await Promise.all(
+      Object.values(body.files).map(({ path }) => unlink(pathToFileURL(path)))
+    );
   }
 }
